fiboooonaaaaacccciii

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -11,7 +11,7 @@ int main(void)
 {
 	unsigned long int a = 1;
 	unsigned long int b = 2;
-	long int c;
+	int c;
 	int fib;
 
 	printf("%lu, ", a);
